refactor: user-service delete-pair return

+++ user-service/src/application/user/command/delete-pair/delete-pair.command-handler.ts
@@ -1,7 +1,9 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { DeletePairCommand } from './delete-pair.command';
-import { NotFoundException } from '@nestjs/common';
+import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { UserRepository } from 'src/domain/user/repository';
+import { ERROR } from 'src/infrastructure/user/common/constant';
+import { UserAggregate } from 'src/domain/user';
 
 @CommandHandler(DeletePairCommand)
 export class DeletePairCommandHandler
@@ -9,7 +11,7 @@ export class DeletePairCommandHandler
 {
   constructor(private readonly repository: UserRepository) {}
 
-  async execute(command: DeletePairCommand): Promise<string> {
+  async execute(command: DeletePairCommand): Promise<UserAggregate> {
     const { userId, pairId } = command;
 
     const pair = await this.repository.findPair(pairId, userId);
@@ -17,8 +19,16 @@ export class DeletePairCommandHandler
       throw new NotFoundException();
     }
 
+    const place = await this.repository.findPlace(userId);
+
+    if (!place || !pair.place) {
+      throw new BadRequestException(ERROR.NULL_PLACE);
+    }
+
+    pair.setDistanceBetweenPlaces(place);
+
     await this.repository.deletePair(pair.id, userId);
 
-    return pair.id;
+    return pair;
   }
 }
